==> 03  -TÉCNICAS DE PROGRAMAÇÃO I/05 - PANDAS/01 - CONTABILIZANDO_NUM_CONCURSO_PUBLICO.ipynb <==
### 01 - Em um concurso público foram contabilizados os números de pessoas inscritas (até a data inicialmente estabelecida para a inscrição) segundo os níveis de escolaridade: fundamental completo, médio completo, superior completo e pós-graduação completa. Segue abaixo a tabela com os valores observados.

```
Nível de escolaridade	Inscritos
Fundamental completo	451
Médio completo	627
Superior completo	292
Pós-graduação completa	95
```
**Com base na tabela acima, crie uma Series com esses valores e calcule:**

* (a) O número total de inscritos
* (b) Número inscritos que possuem, pelo menos, o superior completo.
* (c) Suponha que a inscrição foi prorrogada e, com isso, foi obtido um número adicional de inscrições: 68, 93, 22 e 0 inscritos a mais em cada uma das categorias de escolaridade apresentados na tabela acima (na mesma ordem). Assim, crie uma nova Series com esses valores.
* (d) Utilizando as duas Series que você tem, calcule o número total de inscritos após a prorrogação do período de inscrições.

import pandas as pd

concurso = pd.Series([451, 627, 292, 95],
            index=['Fundamental Completo', 'Médio Incompleto', 'Superior Completo', 'Pós-Graduação Completa'])

# a
print(concurso.sum())
print('-'*50,'\n')

# b
print(concurso['Superior Completo':].sum())
print('-'*50,'\n')

# c
add = pd.Series([68, 93, 22, 0],
               index=['Fundamental Completo', 'Médio Incompleto', 'Superior Completo', 'Pós-Graduação Completa'])
print(add)
print('-'*50,'\n')

# d
concurso_total = concurso + add
print(concurso_total.sum())
