==> 03  -TÉCNICAS DE PROGRAMAÇÃO I/05 - PANDAS/02 - NOTAS_DOS_ALUNOS.ipynb <==
###  02-  Uma turma de cinco alunos foi submetida a uma avaliação e a nota de cada um dos alunos é apresentada na tabela abaixo:

```
Nome do Aluno	Nota
Wilfred	2
Abbie	7
Harry	5.5
Julia	10
Carrie	6.2
```

* (a) Crie uma Series com os valores da tabela acima.
* ((b) Retorne o número de alunos que foram aprovados (considerando que a nota mínima para a aprovação é 7).
* ((c) Qual foi a nota média da turma.
* ((d) Obtenha o nome dos alunos que tiraram notas acima da média da turma.

import pandas as pd

#a
notas = pd.Series([2, 7, 5.5, 10, 6.2],
            index=['Wilfred', 'Abbie' , 'Harry', 'Julia','Carrie'])


# b
print(notas[notas >= 7].count())
print('-'*50,'\n')
# c
print(notas.mean())
print('-'*50,'\n')
# d
print(notas[notas > notas.mean()].index)
print(', '.join(list(notas[notas > notas.mean()].index))) # se preferir
