==> 03  -TÉCNICAS DE PROGRAMAÇÃO I/05 - PANDAS/03 - DATAFRAME_PESO_ALTURA.ipynb <==
### 03 - Crie um objeto DataFrame que contenha os dados da tabela abaixo, os quais se referem ao peso, altura e idade de cinco pessoas:

```
Peso (kg)	Altura (cm)	Idade
72	180	31
80	170	21
60	175	16
90	174	28
100	185	23
```
**Com base no DataFrame criado, realize o que pedido nos itens abaixo:

* (a) Visualize apenas os dados da coluna que contém a altura das pessoas.
* (b) Obtenha a idade média das pessoas.
* (c) Obtenha os dados das pessoas que estão abaixo da altura média.
* (d) Crie um nova coluna no DataFrame que contenha o IMC (Índice de Massa Corporal) de cada pessoa.
* (e) Remova a coluna que contém a idade das pessoas.
* (f) Salve esses dados do DataFrame em um arquivo csv.

import pandas as pd

dados = {'Peso (kg)': [72, 80, 60, 90, 100],
         'Altura (cm)': [180, 170, 175, 174, 185],
         'Idade': [31, 21, 16, 28, 23]}

df = pd.DataFrame(dados)

 # a
df['Altura (cm)']


# b
df['Idade'].mean()

# c
print(df['Altura (cm)'].mean())

df[df['Altura (cm)'] < df['Altura (cm)'].mean()]


 # d

IMC = df['Peso (kg)'] / (df['Altura (cm)'] / 100)**2

df['IMC'] = IMC
df['IMC'] 


 # e

df.drop('Idade', axis=1, inplace=True)

# f

df.to_csv('01-peso-altura-imc.csv', index=False)
