==> 03  -TÉCNICAS DE PROGRAMAÇÃO I/05 - PANDAS/04 -  DEFININDO_NOVA_COLUNA.ipynb <==
### 04 - Carregue, em um objeto DataFrame, o arquivo CSV que você salvou no item (f) da questão anterior. Em seguida,

* (a) Crie uma nova coluna que contenha o nome das pessoas (sinta-se à vontade para atribuir os nomes que preferir).
* (b) Defina a nova coluna criada (com os nomes) como o índice (index) do seu DataFrame.

import pandas as pd

# DADOS DA QUESTÃO ANTERIOR

dados = {'Peso (kg)': [72, 80, 60, 90, 100],
         'Altura (cm)': [180, 170, 175, 174, 185],
         'Idade': [31, 21, 16, 28, 23]}

df = pd.DataFrame(dados)

 # a
df['Altura (cm)']


# b
df['Idade'].mean()


# c
print(df['Altura (cm)'].mean())

df[df['Altura (cm)'] < df['Altura (cm)'].mean()]



 # d

IMC = df['Peso (kg)'] / (df['Altura (cm)'] / 100)**2

df['IMC'] = IMC
df['IMC'] 


 # e

df.drop('Idade', axis=1, inplace=True)


# f

df.to_csv('peso-altura-imc.csv', index=False)

import pandas as pd

df = pd.read_csv('peso-altura-imc.csv')


# a

df['Nome'] = ['Ryu', 'Ken', 'Akuma', 'Chun Li', 'Sagat']


# b

df.set_index('Nome', inplace=True)

 # c

nomes = df[(df['IMC'] >= 25) & (df['IMC'] < 30)].index
list(nomes)
