==> 03  -TÉCNICAS DE PROGRAMAÇÃO I/05 - PANDAS/05 - SALVANDO_DATAFRAMA_EM_xlsx.ipynb <==
### 05 - Carregue, em um objeto DataFrame, o mesmo dataset que você salvou na Questão 3. Em seguida,

* (a) Adicione uma nova coluna nesse DF chamada Classificação, que contenha a classificação de cada indivíduo, de acordo com o seu IMC. Para isso, consulte a tabela abaixo.

```
IMC	Classificação	Obesidade (grau)
Menor que 18,5	Magreza	0
Entre 18,5 e 24,9	Normal	0
Entre 25,0 e 29,9	Sobrepeso	I
Entre 30,0 e 39,9	Obesidade	II
Maior que 40,0	Obesidade Grave	III

```
* (b) Salve em DataFrame no formato xlsx.

# a

import pandas as pd
import numpy as np

df = pd.read_csv('peso-altura-imc.csv')

classificacao = [''] * df.shape[0]

for i in range(df.shape[0]):
    imc = df.loc[i, 'IMC']
    if (imc < 18.5):
        classificacao[i] = 'Abaixo do peso'
    elif (imc < 25):
        classificacao[i] = 'Normal'
    elif (imc < 30):
        classificacao[i] = 'Sobrepeso'
    elif imc < 40:
        classificacao[i] = 'Obesidade'
    else:
        classificacao[i] = 'Obesidade Grave'

df['Classificação'] = classificacao

# b

df.to_excel('peso-altura-imc.xlsx', index=False)
