==> 03  -TÉCNICAS DE PROGRAMAÇÃO I/05 - PANDAS/06 -  JUTANDO_DADOS_DE_2_ARQUIVOS.ipynb <==
### 06 - O arquivo avocado.csv consiste em um dataset que contém dados sobre vendas de abacates (avocado, em inglês) em diversas regiões dos Estados Unidos. Essa dataset contém as seguintes colunas:

* Date - a data da observação
* AveragePrice - o preço médio de um único abacate
* year - o ano
* region - a cidade ou região da observação
* Total Volume - volume total de abacates vendidos

Além das informações acima, contidas no arquivo avocado.csv, o tipo do abacate (convencional ou orgânico) também seria uma informação importante para a sua análise. Suponhamos que você conseguiu essas informações para cada uma das observações do arquivo avocado.csv, na mesma ordem, e salvou-as no arquivo avocado_type.csv. Portanto, tudo o que você precisa é juntar esses dois arquivos em um mesmo DataFrame. E, em seguida, salve-o em um arquivo csv.

**PS. Caso queira criar uma Series com os dados aleatórios para o tipo de abacate, utilize o código abaixo:**


```
tipos = pd.Series(np.random.randint(0, 2, df_avocado.shape[0]))
tipos = tipos.map({0: 'convencional', 1:'orgânico'})
```

import pandas as pd

df_avocado = pd.read_csv('avocado.csv')
df_type = pd.read_csv('avocado_type.csv')

new_df = pd.concat([df_avocado, df_type], axis=1)

new_df.to_csv('avocado-full.csv', index=False)
