==> 03  -TÉCNICAS DE PROGRAMAÇÃO I/05 - PANDAS/07 - ANALISANDO_QTD_DE_DADOS.ipynb <==
### 07 - Carregue o arquivo csv que você salvou na questão anterior e, em seguida, obtenha o que é solicitado nos itens elencados abaixo:

* (a) A quantidade de abacates do tipo convecional e do tipo orgânico.
* (b) A quantidade de registros (linhas) de abacates vendidos para cada região.
* (c) Para que todos os labels das colunas fiquem padronizados, renomeie aqueles que estão iniciando em letra minúscula para que iniciem em letra maiúscula.
* (d) Uma vez que você já possui uma coluna com a data, remova a coluna que apresenta o ano.

import pandas as pd

df = pd.read_csv('avocado-full.csv')

df.head()

# a)
df['type'].value_counts()


# b)
print(df['region'].value_counts())


# c)
df.rename(columns={'year': 'Year', 'region': 'Region', 'type': 'Type'}, inplace=True)

# d)
df.drop('Year', axis=1, inplace=True)
