==> 03  -TÉCNICAS DE PROGRAMAÇÃO I/05 - PANDAS/08 - AGRUPANDO_DADOS.ipynb <==
### 03 Carregue, em um objeto DataFrame, o mesmo dataset que você salvou na questão Q1, agrupe os dados do DF pelo tipo do abacate e, em seguida, obtenha:

* (a) A volume total de abacates vendidos para cada tipo.
* (b) O preço médio dos abacates vendidos para cada tipo.

import pandas as pd

df = pd.read_csv('avocado-full.csv')

df_type = df.groupby('type')


# a)
df_type.sum()['Total Volume']


# b)
df_type.mean()['AveragePrice']
