==> 03  -TÉCNICAS DE PROGRAMAÇÃO I/05 - PANDAS/09 - VERIFICANDO_DADOS_DUPLICADOS.ipynb <==
### 04 - O arquivo googleplaystore.csv é um dataset que contém diversas informações sobre aplicativos disponíveis para download na Google Play Store. Logo após, verifique se existem dados duplicados e/ou ausentes nesse dataset.

### Caso existam dados duplicados, remova-os. Caso existam dados ausentes, você pode removê-los ou pensar em alguma maneira de preenchê-los.

import pandas as pd

df = pd.read_csv('googleplaystore.csv')

df.head()

df.drop_duplicates(inplace=True)

# Percentual de dados ausentes na coluna Rating
df['Rating'].isnull().sum() / df.shape[0] # 1465 / 10358

df['Rating'].fillna(df['Rating'].mean(), inplace=True)

df['Type'].fillna('Free', inplace=True)

# Difícil prever a versão do Aplicativo e do Android, portanto...
df.dropna(inplace=True)
