==> 03  -TÉCNICAS DE PROGRAMAÇÃO I/05 - PANDAS/11 - CRIANDO_SERIES_TOP10.ipynb <==
### 11 - Crie uma Series com os índices registrados pelo TIOBE Index para as top 10 linguagens de programação. Os índices da Series devem ser os nomes das linguagens, e os valores devem ser as taxas de popularidade de cada respectiva linguagem de programação. Em seguida, utilize os métodos da Series para:

* Exibir os cinco primeiros valores da Series
( Exibir os cinco últimos valores da Series

import pandas as pd

# Dados de agosto de 2021
index = [
    'C',
    'Python',
    'Java',
    'C++',
    'C#',
    'Visual Basic',
    'JavaScript',
    'PHP',
    'Assembly Language',
    'SQL'
]

taxas = [12.57, 11.86, 10.43, 7.36, 5.14, 4.67, 2.95, 2.19, 2.03, 1.47]

tiobe_series = pd.Series(taxas, index=index)

print(tiobe_series.iloc[:5])
print(tiobe_series.iloc[5:])
