==> 03  -TÉCNICAS DE PROGRAMAÇÃO I/05 - PANDAS/12 - SERIES_ORDEM_ALFABETICA.ipynb <==
### 12- Crie uma Series com os 11 primeiros termos da sequência de Fibonacci. Utilize letras (em ordem alfabética) como os índices da Series. Em seguida, obtenha apenas aqueles elementos que forem superiores a 10 e inferiores a 50.

```Sequência de Fibonacci: 0, 1, 1, 2, 3, 5, 8, 13, 21, 34, 55...```

import pandas as pd
import numpy as np

# Gerando dados aleatórios
numeros = np.random.randint(1, 6, 20)
print(numeros)

series = pd.Series(numeros)

# Pode haver mais de uma moda, mas escolheremos a primeira
mais_frequente = series.mode()[0]
print(f'Valor mais frequente: {mais_frequente}')

# Substituição
series[series != mais_frequente] = "Outro valor"

print(series)
